[HDRP] Fix debug modes and optim SSR
Fix async SSR that have an uneeded barrier on xbox for reasons
https://unity.slack.com/archives/C6Y79CZM0/p1716461545320749?thread_ts=1715329897.499809&cid=C6Y79CZM0

Fix unecessary texture allocation when ssr approximate mode is on
Fix unecessary rendergraph barrier on coarse stencil (code is disabled with a comment on shader)
Fix triple upload of constant buffer with same data
Also fix various debug modes that got broken by fog

Trunk:
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/c5cdea7d-7af4-47d4-92a1-a63afbc39a50)

Fix:
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/fad6f1e6-513e-43dd-a289-1fc16a770986)

+++ Packages/com.unity.render-pipelines.high-definition/Runtime/Lighting/LightDefinition.cs.hlsl
@@ -83,7 +83,7 @@ struct CelestialBodyData
     float3 flareColor;
     float flareFalloff;
     float3 padding;
-    float shadowIndex;
+    int shadowIndex;
 };
 
 // Generated from UnityEngine.Rendering.HighDefinition.DirectionalLightData
